feat: add metadata for client component
